Flask/requirements.txt
Flask==2.3.3
pymongo==4.5.0
textblob==0.17.1
dnspython==2.3.0
python-dotenv==1.0.0
aiohttp==3.8.5
requests==2.31.0
